Refactor location of the token service
