+++ src/main/java/de/fearnixx/jeak/service/controller/token/ITokenService.java
@@ -1,4 +1,4 @@
-package de.fearnixx.jeak.service.controller.connection;
+package de.fearnixx.jeak.service.controller.token;
 
 public interface ITokenService {
     boolean verifyToken(Class<?> controllerClass, String token);
